% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_data.R
\name{df_norm_0_1}
\alias{df_norm_0_1}
\title{df_norm_0_1(df, column = 1)}
\usage{
df_norm_0_1(df, column = 1)
}
\arguments{
\item{df}{data frame name}

\item{column}{column numeric specification which will be normalized}
}
\value{
data frame with normalize column
}
\description{
normalize data_frame column to interval [0,1]
}
\examples{
          df <- data.frame(x= 1:10, y = 21:30)
          df_norm_0_1(df, column = 2)
}
