% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dataframe_util.R
\name{lst2df}
\alias{lst2df}
\title{List of data.frames to single data.frame conversion}
\usage{
lst2df(lst, nms = names(lst))
}
\arguments{
\item{lst}{list of data.frames for conversion}

\item{nms}{names of data.frames that will be added as additional column
By default sets names(lst)}
}
\value{
joined data.frame with additional column
}
\description{
conversion of list of data.frames to single data frame.
             Can be added aditional column with data.frame names
}
\examples{
 df1 <- data.frame(a = c(0,1,2), b = c(10,11,12))
 df2 <- data.frame(a = c(110,111,112), b = c(1110,1111,1112))
 lst <- list()
 lst[[1]] <- df1
 lst[[2]] <- df2
 names(lst) <- c("DF1", "DF2")
 lst2df(lst, names(lst))
}
\author{
Krunoslav Juraic
}

