% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dataframe_util.R
\name{plot_tables}
\alias{plot_tables}
\title{preview plot of single data.frame in list}
\usage{
plot_tables(lst, col_x = 1, col_y = 2, ...)
}
\arguments{
\item{lst}{list of data.frames for preview}

\item{col_x}{x values for plot}

\item{col_y}{y values for plot}

\item{...}{additional parameters transfered to plot function}
}
\value{
None
}
\description{
Use manipulate to preview plot data from list of data.frames.
             (X,Y) pair of column shoud be specified. Additional graphic
             parametrs for plot function can be specified.
             Usually this function is used in combination wity function
             read_tables().
             As plot main title it's used data.frame name.
}
\examples{
   \dontrun{
            dat <- read_tables()
            plot_tables(dat, col = 2)
   }
}
\author{
K. Juraic
}

