% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dataframe_util.R
\name{read_tables}
\alias{read_tables}
\title{read multiple tabular files to list of data.frames}
\usage{
read_tables(fnms = tk_choose.files(), ...)
}
\arguments{
\item{fnms}{list of file names}

\item{...}{additional parameters for read.table function}
}
\value{
list of data.frames with readed data
}
\description{
Read tabular data from multiple files in list of data.frames
}
\details{
Function is based on read.table function. If the list of file.name
         is not set it's open tk_file_dialog to set multiple files with data.
         It's using read.table function. Additional parameters for read.table
         function can be simply added to function call like skip, header.
         Return list of data.frames. Names of list elements set to base file
         names with data.
}
\examples{
       \dontrun{read_tables(skip=26, header=TRUE)}
}
\author{
K. Juraic
}
